+++ test/conftest.py
@@ -4,17 +4,19 @@ from app import create_app
 from .test_models import create_factory
 
 
-@pytest.fixture
-def app(mocker):
-    mocker.patch("flask_sqlalchemy.SQLAlchemy.init_app", return_value=True)
-    mocker.patch("flask_sqlalchemy.SQLAlchemy.create_all", return_value=True)
-    mocker.patch("api.routes.products", return_value=[])
-    mocker.patch("api.routes.product_detail", return_value={"id": 1})
+create_factory()
 
-    create_factory()
+
+@pytest.fixture
+def app():
+    # mocker.patch("flask_sqlalchemy.SQLAlchemy.init_app", return_value=True)
+    # mocker.patch("flask_sqlalchemy.SQLAlchemy.create_all", return_value=True)
+    # mocker.patch("api.routes.products", return_value=[])
+    # mocker.patch("api.routes.product_detail", return_value={"id": 1})
     app = create_app()
     return app
 
+   
 
 # import pytest
