+++ test/settings.py
@@ -1,0 +1,1 @@
+DEFAULT_PRICE = 1000
