+++ test/test_routes.py
@@ -1,19 +1,13 @@
-# def test_example(client):
-#     response = client.get("/api/products/1")
-#     print('Response', response)
-#     assert response.status_code == 200
+import json
 
-# from flask import session
+def test_product_detail_api(client):
+    id = 1
+    response = client.get(f"/api/products/{id}")
+    assert response.status_code == 200
+    data = json.loads(response.data)
+    assert id == data.get('id')
 
-# def test_access_session(app):
-#     c = app.test_client()
-#     response = c.get('/api/products/1')
+def test_product_api(client):
+    response = client.get("/api/products")
+    assert response.status_code == 200
 
-# from unittest.mock import MagicMock, patch
-
-# from api.routes import products
-
-# def test_LegendsPostService_can_init():
-#     session = MagicMock()
-#     service = products(_session=session)
-#     assert service.session is session
